+++ com.microej.example.iot.ssl.mqtt.publisher/src/main/java/com/microej/example/iot/mqtt/HelloWorldConstants.java
@@ -6,10 +6,13 @@
  */
 package com.microej.example.iot.mqtt;
 
+import java.util.Random;
+
 public interface HelloWorldConstants {
 
 	String BROKER = "ssl://test.mosquitto.org:8883";
-	String PUBLISHER_ID = "publisher";
-	String HELLO_WORLD_MESSAGE = "Hello World !!";
-	String TOPIC = "microej";
+	// Set a random id.
+	String PUBLISHER_ID = "publisher_" + new Random().nextInt(); //$NON-NLS-1$
+	String HELLO_WORLD_MESSAGE = "Hello World !!"; //$NON-NLS-1$
+	String TOPIC = "microej"; //$NON-NLS-1$
 }