+++ com.microej.example.iot.ssl.restget/src/main/java/com/microej/example/iot/ssl/restget/GetBS.java
@@ -1,7 +1,7 @@
 /*
  * Java
  *
- * Copyright 2018 IS2T. All rights reserved.
+ * Copyright 2015-2018 IS2T. All rights reserved.
  * For demonstration purpose only.
  * IS2T PROPRIETARY. Use is subject to license terms.
  */
@@ -13,6 +13,7 @@ import java.net.HttpURLConnection;
 import java.security.KeyStore;
 import java.security.cert.Certificate;
 import java.security.cert.CertificateFactory;
+import java.util.Calendar;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -23,47 +24,36 @@ import javax.net.ssl.TrustManagerFactory;
 
 import org.json.me.JSONArray;
 
+import android.net.SntpClient;
+import ej.bon.Util;
 import ej.rest.web.JSONResource;
 import ej.rest.web.Resty;
-import ej.wadapps.app.Activity;
+import ej.wadapps.app.BackgroundService;
 
 /**
  *
  *
  */
-public class GetActivity implements Activity {
+public class GetBS implements BackgroundService {
 
-	public static final Logger LOGGER = java.util.logging.Logger.getLogger("HTTPS GET");
+	private static final Logger LOGGER = java.util.logging.Logger.getLogger("HTTPS GET"); //$NON-NLS-1$
 
 	// The server certificate file name
-	public static String SERVER_CERT_FILENAME = "communitystore.microej.com.crt";
-	public static String SERVER_CERT_PATH = "/certificates/";
+	private static final String SERVER_CERT_FILENAME = "GlobalSignRootCA.crt"; //$NON-NLS-1$
+	private static final String SERVER_CERT_PATH = "/certificates/"; //$NON-NLS-1$
 
 	// X509 certificate type name
-	public static String CERT_TYPE = "X509";
+	private static final String CERT_TYPE = "X509"; //$NON-NLS-1$
 
 	// TLS algorithm version 1.2
-	public static String TLS_VERSION_1_2 = "TLSv1.2";
+	private static final String TLS_VERSION_1_2 = "TLSv1.2"; //$NON-NLS-1$
 
 	// The server url
-	public static String SERVER_URL = "https://preprodstore.microej.com";
+	private static final String SERVER_URL = "https://communitystore.microej.com"; //$NON-NLS-1$
 
-	@Override
-	public String getID() {
-		// TODO Auto-generated method stub
-		return null;
-	}
-
-	@Override
-	public void onCreate() {
-		// TODO Auto-generated method stub
-
-	}
-
-	@Override
-	public void onRestart() {
-		// TODO Auto-generated method stub
-
+	public static void main(String[] args) {
+		updateTime();
+		new GetBS().onStart();
 	}
 
 	@Override
@@ -73,9 +63,9 @@ public class GetActivity implements Activity {
 
 		try {
 
-			GetActivity.initRestyHttpsContext();
+			GetBS.initRestyHttpsContext();
 
-			GetActivity.doGetRequest();
+			GetBS.doGetRequest();
 
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -84,27 +74,8 @@ public class GetActivity implements Activity {
 	}
 
 	@Override
-	public void onResume() {
-		// TODO Auto-generated method stub
-
-	}
-
-	@Override
-	public void onPause() {
-		// TODO Auto-generated method stub
-
-	}
-
-	@Override
 	public void onStop() {
-		// TODO Auto-generated method stub
-
-	}
-
-	@Override
-	public void onDestroy() {
-		// TODO Auto-generated method stub
-
+		// Nothing to do.
 	}
 
 	/**
@@ -117,12 +88,10 @@ public class GetActivity implements Activity {
 		 * Create and initialize the SSLContext which will be used to connect to the secure Server. The followings steps
 		 * show how to create and setup the SSLContext for Resty Https connection.
 		 */
-		try (
-				/*
-				 * Step 1 : Create an input stream with the server certificate file
-				 */
-
-				InputStream in = GetActivity.class.getResourceAsStream(SERVER_CERT_PATH + SERVER_CERT_FILENAME)) {
+		/*
+		 * Step 1 : Create an input stream with the server certificate file
+		 */
+		try (InputStream in = GetBS.class.getResourceAsStream(SERVER_CERT_PATH + SERVER_CERT_FILENAME)) {
 
 			/*
 			 * Step 2 : Generate the server certificate
@@ -140,7 +109,7 @@ public class GetActivity implements Activity {
 			// parameters
 			store.load(null, null);
 			// add the server certificate to our created KeyStore
-			store.setCertificateEntry("myServer", myServerCert);
+			store.setCertificateEntry("myServer", myServerCert); //$NON-NLS-1$
 
 			/*
 			 * Step 4: Create and initialize the trust manager with our KeyStore
@@ -173,8 +142,8 @@ public class GetActivity implements Activity {
 	 */
 	public static void doGetRequest() throws Exception {
 
-		LOGGER.info("=========== GET REQUEST ===========");
-		String requestURL = SERVER_URL + "/api/v2/DeviceReferences";
+		LOGGER.info("=========== GET REQUEST ==========="); //$NON-NLS-1$
+		String requestURL = SERVER_URL + "/api/v2/DeviceReferences"; //$NON-NLS-1$
 		Resty resty = new Resty();
 
 		// do GET request request;
@@ -189,13 +158,40 @@ public class GetActivity implements Activity {
 				LOGGER.info(response.toString(2));
 
 			} else {
-				throw new IOException("Wrong response code " + responseCode + " for GET " + requestURL);
+				throw new IOException("Wrong response code " + responseCode + " for GET " + requestURL); //$NON-NLS-1$ //$NON-NLS-2$
 			}
 
 		} finally {
 			conn.disconnect();
 		}
 
+	}
+
+	/**
+	 *
+	 */
+	public static void updateTime() {
+		LOGGER.info("=========== Updating time ==========="); //$NON-NLS-1$
+		SntpClient ntpClient = new SntpClient();
+
+		while (Util.currentTimeMillis() < 1000000) {
+			/**
+			 * Request NTP time
+			 */
+			if (ntpClient.requestTime("ntp.ubuntu.com", 123, 1000)) { //$NON-NLS-1$
+				long now = ntpClient.getNtpTime() + Util.platformTimeMillis() - ntpClient.getNtpTimeReference();
+
+				Calendar.getInstance().setTimeInMillis(now);
+				Util.setCurrentTimeMillis(now);
+			} else {
+				try {
+					Thread.sleep(1000);
+				} catch (InterruptedException e) {
+					// Sanity.
+				}
+			}
+		}
+		LOGGER.info("Time updated"); //$NON-NLS-1$
 	}
 
 }