+++ com.microej.example.iot.androidconnectivity/CHANGELOG.md
@@ -1,8 +1,21 @@
+## 2.1.0 (2018-10-09)
+
+Features:
+  - Support android connectivity 1.1.0.
+  - Fix minor version on ej.api.
+
 ## 2.0.0 (2016-12-29)
+
 Features:
   - Change organization.
   
 ## 1.0.0 (2016-06-13)
+
 Features:
   - Initial revision.
 
+<!--
+  Copyright 2016-2018 IS2T. All rights reserved.
+  For demonstration purpose only.
+  IS2T PROPRIETARY. Use is subject to license terms.
+-->