+++ com.microej.example.iot.json/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.1 (2018-10-09)
+
+Features:
+  - Fix minor version on ej.api.
+
 ## 3.0.0 (2016-12-29)
 Features:
   - Change organization.
@@ -12,3 +17,9 @@ Port to MicroEj 4
 Features:
   - Initial revision.
 
+
+<!--
+  Copyright 2016-2018 IS2T. All rights reserved.
+  For demonstration purpose only.
+  IS2T PROPRIETARY. Use is subject to license terms.
+-->